+++ packages/dartness_generator/lib/builder.dart
@@ -4,6 +4,7 @@ import 'package:source_gen/source_gen.dart';
 import 'src/application_generator.dart';
 import 'src/controller_generator.dart';
 import 'src/error_handler_generator.dart';
+import 'src/scheduler_generator.dart';
 
 Builder controller(BuilderOptions options) =>
     SharedPartBuilder([ControllerGenerator()], 'controller');
@@ -13,3 +14,6 @@ Builder errorHandler(BuilderOptions options) =>
 
 Builder application(BuilderOptions options) =>
     SharedPartBuilder([ApplicationGenerator()], 'application');
+
+Builder scheduler(BuilderOptions options) =>
+    SharedPartBuilder([SchedulerGenerator()], 'scheduler');