+++ packages/dartness_generator/lib/src/error_handler_generator.dart
@@ -83,9 +83,9 @@ class ErrorHandlerGenerator extends GeneratorForAnnotation<ErrorHandler> {
 
     final errorTypes = catchErrorAnnotation?.getField('errors');
     final errors = errorTypes?.toListValue()?.map((errorType) {
-          final typeName = errorType
-                  .toTypeValue()
-                  ?.getDisplayString(withNullability: false) ??
+          final typeName = errorType.toTypeValue()?.getDisplayString(
+                    withNullability: false,
+                  ) ??
               '';
           final typeCode = Code(typeName);
           return CodeExpression(typeCode);