+++ packages/dartness_server/lib/schedule.dart
@@ -2,6 +2,6 @@
 library schedule;
 
 export 'src/configuration/schedule/scheduled.dart';
+export 'src/configuration/schedule/scheduler.dart';
 export 'src/configuration/schedule/time_unit.dart';
-export 'src/configuration/schedule/cron_expression.dart';
-export 'src/configuration/schedule/cron_scheduler.dart';
+export 'src/configuration/schedule/scheduler_manager.dart';