+++ packages/dartness_server/lib/src/configuration/schedule/scheduler.dart
@@ -1,0 +1,4 @@
+/// Represents an annotation for classes with scheduling tasks.
+class Scheduler {
+  const Scheduler();
+}